% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Cbeta.R
\name{Calc_Cbeta}
\alias{Calc_Cbeta}
\title{Calc_Cbeta}
\usage{
Calc_Cbeta(beta)
}
\arguments{
\item{beta}{The dataframe storing the magnitude of the jumps beta of the coefficients of the additive hazard.  The first column should be the survival time.}
}
\value{
Return the cumulative beta by the giving beta.
}
\description{
Computing cumulative beta from beta (the magnitude of the jumps of the coefficient function)
}
\examples{
beta  = data.frame(time = c(1,2,3,4), beta = c(5,6,7,8))
cbeta = Calc_Cbeta(beta)
}
